requirements.txt
asgiref==3.5.2
Django==3.2
django-filter==22.1
djangorestframework==3.14.0
gunicorn==20.1.0
pytz==2022.2.1
sqlparse==0.4.3
whitenoise==6.2.0
